# Improve performance of SetTags