While debugging slowness in `groupByTags`, the performance of SetTags stood out. I don't believe this was the primary source of slowness in my case, but it certainly could be a contributing factor, since this is called for every series that is processed via `render`.  Aside from replacing `SplitN` with `IndexByte`, I also only allocate a new map if the existing one is `nil`. I ran the benchmark in two modes, one where I leave the Tags and one where I `nil` them out.

Here's the benchcmp when `s.Tags == nil`
```
benchmark                             old ns/op     new ns/op     delta
BenchmarkSetTags_00tags_00chars-8     272           193           -29.04%
BenchmarkSetTags_20tags_32chars-8     3725          1798          -51.73%

benchmark                             old MB/s     new MB/s     speedup
BenchmarkSetTags_00tags_00chars-8     51.36        72.48        1.41x
BenchmarkSetTags_20tags_32chars-8     358.05       741.88       2.07x

benchmark                             old allocs     new allocs     delta
BenchmarkSetTags_00tags_00chars-8     3              2              -33.33%
BenchmarkSetTags_20tags_32chars-8     23             2              -91.30%

benchmark                             old bytes     new bytes     delta
BenchmarkSetTags_00tags_00chars-8     352           336           -4.55%
BenchmarkSetTags_20tags_32chars-8     2256          1264          -43.97%
```

and here is when we reuse the Tags:
```
benchmark                             old ns/op     new ns/op     delta
BenchmarkSetTags_00tags_00chars-8     272           99.3          -63.49%
BenchmarkSetTags_20tags_32chars-8     3725          2081          -44.13%

benchmark                             old MB/s     new MB/s     speedup
BenchmarkSetTags_00tags_00chars-8     51.36        140.92       2.74x
BenchmarkSetTags_20tags_32chars-8     358.05       640.78       1.79x

benchmark                             old allocs     new allocs     delta
BenchmarkSetTags_00tags_00chars-8     3              0              -100.00%
BenchmarkSetTags_20tags_32chars-8     23             0              -100.00%

benchmark                             old bytes     new bytes     delta
BenchmarkSetTags_00tags_00chars-8     352           0             -100.00%
BenchmarkSetTags_20tags_32chars-8     2256          0             -100.00%
```

With a lot of tags, the cost of clearing the map is higher than reallocating, but this was go 1.10.3 and I know go 1.11 has optimized map clearing.