@Dieterbe 2018-12-17
> With a lot of tags, the cost of clearing the map is higher than reallocating, but this was go 1.10.3 and I know go 1.11 has optimized map clearing.

correct.  https://go-review.googlesource.com/c/go/+/110055 should pay off here.