@Dieterbe 2018-12-17
>  I ran the benchmark in two modes, one where I leave the Tags and one where I nil them out.

did you do this by uncommenting the nil setting line? instead of that, perhaps we should just have 2 separate benchmarks? to simplify reproducing the stats