@shanson7 2018-12-17
Upgraded my local golang to 1.11.4 and now the map clearing approach shows it's benefit:
```
BenchmarkSetTags_20tags_32chars-8          	10000000	      1666 ns/op	 800.46 MB/s	    1264 B/op	       2 allocs/op
BenchmarkSetTags_20tags_32chars_reused-8   	20000000	      1096 ns/op	1216.87 MB/s	     288 B/op	       1 allocs/op
```